Refactor always folders in modules
