+++ src/app/conceptor/conceptor-routing.module_LOCAL_912.ts
@@ -2,7 +2,7 @@ import { Component, NgModule } from '@angular/core';
 import { OutletContext, RouterModule, Routes } from '@angular/router';
 import { SidebarComponent } from './sidebar/sidebar.component';
 
-import { AddModuleComponent } from './add-module/add-module.component';
+import { AddModuleComponent } from './modules/add-module/add-module.component';
 
 import { ListComponent } from '../course/list/list.component';
