+++ src/app/conceptor/conceptor.module.ts
@@ -5,7 +5,7 @@ import { ConceptorRoutingModule } from './conceptor-routing.module';
 import { SharedModule } from '../shared/shared.module';
 import { UiModule } from '../ui/ui.module';
 import { SidebarComponent } from './sidebar/sidebar.component';
-import { AddModuleComponent } from './add-module/add-module.component';
+import { AddModuleComponent } from './modules/add-module/add-module.component';
 import { CourseModule } from '../course/course.module';
 import { CourseHandlerComponent } from '../course/course-handler/course-handler.component';
 import { MediaDialogComponent } from './modules/media-dialog/media-dialog.component';
