+++ src/app/conceptor/modules/media-dialog/media-dialog.component.scss
@@ -1,0 +1,5 @@
+.list-media{
+    background-color:rgb(111, 111, 229);
+    border-radius: 10px;
+
+}
