+++ src/app/conceptor/modules/media-dialog/media-dialog.component.ts
@@ -1,10 +1,9 @@
 import { Component, Inject, OnInit } from '@angular/core';
-import { MediaService } from '../../services/media.service';
+import { MediaService } from '../services/media.service';
 import { MediaType } from 'src/app/course/types/media-type';
 import { take } from 'rxjs';
-import { ModuleType } from '../../types/module-type';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { AddModuleComponent } from '../../add-module/add-module.component';
+import { AddModuleComponent } from '../add-module/add-module.component';
 
 @Component({
   selector: 'app-media-dialog',
@@ -19,7 +18,8 @@ export class MediaDialogComponent implements OnInit {
   constructor(
     private _mediaService: MediaService,
     //public dialogRef: MatDialogRef<AddModuleComponent>,
-    //@Inject(MAT_DIALOG_DATA) public data:any,
+    //@Inject(MAT_DIALOG_DATA) public data: any,
+    
   ) { }
 
   ngOnInit(): void {
@@ -31,9 +31,4 @@ export class MediaDialogComponent implements OnInit {
       this.medias = response
     })
   }
-  /*public addMedia(media: MediaType): void {
-    this.media = media
-    this.dialogRef.close(this.media)
-  }*/
-
 }
