+++ src/app/conceptor/modules/services/module.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import { ModuleType } from '../types/module-type';
+import { ModuleType } from '../../types/module-type';
 
 @Injectable({
   providedIn: 'root'
