+++ src/app/conceptor/sidebar/sidebar.component.html
@@ -67,4 +67,3 @@
   </mat-sidenav-content>
 
 </mat-sidenav-container>
-<app-media-dialog></app-media-dialog>
